refactor: move some funcs from main, reduce data products, change styles, fix some errors
