+++ src/components/copyBtn/copyBtn.scss
@@ -1,6 +1,6 @@
 .copy-btn {
   width: 100px;
-
+  padding: 5px;
   &.copied {
     background-color: greenyellow;
   }
