+++ src/components/filters/filters.scss
@@ -9,3 +9,12 @@
   padding-bottom: 20px;
 }
 
+.checkboxItem {
+  display: flex;
+  padding-right: 10px;
+  margin-bottom: 10px;
+}
+
+.checkboxItem span {
+  margin-left: auto;
+}
