+++ src/components/found/found.scss
@@ -1,0 +1,3 @@
+.found {
+  min-width: 100px;
+}
