+++ src/components/found/found.ts
@@ -1,4 +1,5 @@
 import { BaseComponent } from '../baseComponent';
+import './found.scss';
 
 export class Found extends BaseComponent {
   count = 100;
