+++ src/components/header/header.scss
@@ -45,4 +45,5 @@
 
 .header__total-price {
   min-width: 200px;
+  font-size: 1.1rem;
 }
