+++ src/components/resetBtn/resetBtn.scss
@@ -1,0 +1,4 @@
+.reset-btn {
+  width: 100px;
+  padding: 5px;
+}
