+++ src/components/resetBtn/resetBtn.ts
@@ -1,8 +1,9 @@
 import { BaseComponent } from '../baseComponent';
+import './resetBtn.scss';
 
 export class ResetBtn extends BaseComponent {
   constructor(callback: () => void) {
-    super('button', 'reset-btn');
+    super('button', 'btn reset-btn');
     this.element.textContent = 'Reset Filters';
     this.element.onclick = () => {
       callback();
